+++ lib/src/theme/themes/default_theme_no_secondary_border_variant.dart
@@ -369,10 +369,10 @@ class ShadDefaultThemeNoSecondaryBorderVariant extends ShadThemeVariant {
           width: 1,
         ),
       ),
-      popoverReverseDuration: Duration.zero,
       optionsPadding: const EdgeInsets.all(4),
       showScrollToTopChevron: true,
       showScrollToBottomChevron: true,
+      popoverReverseDuration: Duration.zero,
       anchor: const ShadAnchorAuto(offset: Offset(1, 4)),
       searchPadding: const EdgeInsets.all(12),
     );